requirements.txt
pandas==3.0.1
numpy==2.4.3
scikit-learn==1.6.1
streamlit==1.45.1
langchain==0.3.25
langchain-community==0.3.24
langchain-huggingface==0.1.2
faiss-cpu==1.11.0
sentence-transformers==4.1.0
textblob==0.19.0
torchvision==0.22.0
huggingface-hub==0.32.3
python-dotenv==1.1.0
